implement event expiration protections (and allow admin overrides); pass auth'd user access level back to frontend via header

+++ src/main/java/com/crowdease/yasss/api/RemoveEventEndpoint.java
@@ -14,6 +14,7 @@ import com.axonibyte.lib.http.APIVersion;
 import com.axonibyte.lib.http.rest.EndpointException;
 import com.axonibyte.lib.http.rest.HTTPMethod;
 import com.crowdease.yasss.model.Event;
+import com.crowdease.yasss.model.User.AccessLevel;
 
 import org.json.JSONObject;
 
@@ -52,6 +53,9 @@ public final class RemoveEventEndpoint extends APIEndpoint {
 
       if(!auth.atLeast(event))
         throw new EndpointException(req, "access denied", 403);
+
+      if(!auth.atLeast(AccessLevel.ADMIN) && event.isExpired())
+        throw new EndpointException(req, "event expired", 412);
       
       event.delete();
